refactor : 컴포넌트, 컨텍스트, 타입 분리
